fix: Moved images to page bundles
The publishing pipeline ignores files in the static directory, so this
commit changes all posts that were previously an .md page and static
image into a page bundle. This allows the pipelines to modify the
images and generate optimized versions.

Note: Drafts are included even though they won't be published.